Add deltaToGlue function
the deltaToGlue function takes a S3 Path, which contains a delta log, creates a
manifest out of the log and then creates a new glue partition. This allows
spectrum and athena to query the same S3 tables, and also take advantage of the
fact that delta can optimise the table. Spectrum and Athena tables benefit from
the optimize too.

Add deltaToGlue function

the deltaToGlue function takes a S3 Path, which contains a delta log, creates a
manifest out of the log and then creates a new glue partition. This allows
spectrum and athena to query the same S3 tables, and also take advantage of the
fact that delta can optimise the table. Spectrum and Athena tables benefit from
the optimize too.